# docs/swagger.yaml
basePath: /v1/core-services
definitions:
...
  internal_adapter_handler.StandardErrorEnvelope:
    properties:
      error: {}
      meta: {}
      success:
        type: boolean
    type: object
  internal_adapter_handler.StandardSuccessEnvelope:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'Core service: catalog, orders, master data, and seller APIs for the
    Modami marketplace.'
  title: Modami Core Service API
  version: "1.0"
paths:
...
  /search:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Category filter
        in: query
        name: category_id
        type: string
      - description: Condition
        in: query
        name: condition
        type: string
      - description: Brand
        in: query
        name: brand
        type: string
      - description: Min price
        in: query
        name: min_price
        type: integer
      - description: Max price
        in: query
        name: max_price
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Search products (alias of GET /products/search)
      tags:
      - Search
  /search/suggest:
    get:
      parameters:
      - description: Prefix
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Hashtag suggestions (alias of GET /hashtags/suggest)
      tags:
      - Search
  /search/trending:
    get:
      parameters:
      - description: Max tags
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Trending hashtags (alias of GET /hashtags/trending)
      tags:
      - Search
  /sellers/{id}:
    get:
      parameters:
      - description: Seller user ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Public seller profile
      tags:
      - Sellers
  /sellers/{id}/products:
    get:
      parameters:
      - description: Seller user ID
        in: path
        name: id
        required: true
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Seller's listed products
      tags:
      - Sellers
  /sellers/{id}/reviews:
    get:
      parameters:
      - description: Seller user ID
        in: path
        name: id
        required: true
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Seller reviews
      tags:
      - Sellers
  /sellers/{id}/stats:
    get:
      parameters:
      - description: Seller user ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Seller public stats
      tags:
      - Sellers
  /sellers/{seller_id}/products:
    get:
      parameters:
      - description: Seller user ID
        in: path
        name: seller_id
        required: true
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: List products by seller
      tags:
      - Products
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"